It turns out that regular digitalRead() calls are too slow and bring the arduino down when I use them in the interrupt routines when the motor runs at full speed which results in more than 40000 ticks per second per motor. Now I am using the digitalWriteFast library for high performance reads and writes by jrraines see http://www.arduino.cc/cgi-bin/yabb2/YaBB.pl?num=1267553811/0 and http://code.google.com/p/digitalwritefast/
Fast reads only work when the pins are known at compile time. Hence I needed to abandon the QuadratureEncoder class and bring the encoder code directly into Robot.pde.
The frequency with which the updates in loop() are performed is now cleanly fixed to 50Hz (required addition to TimeInfo.h).
The PID code in SpeedController.h works now.